chore: Rename existing Meganav as LegacyMeganav
to pay wave to the new meganav component
