+++ src/core/LegacyMeganav/component.js
