+++ src/core/LegacyMeganav/component.json
