+++ src/core/MeganavItemsDesktop.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import MeganavData from "./Meganav/component.json";
+import MeganavData from "./LegacyMeganav/component.json";
 import MeganavControl from "./MeganavControl";
-import { AbsUrl, MeganavPanels, MeganavPaths, MeganavTheme } from "./Meganav";
+import { AbsUrl, MeganavPanels, MeganavPaths, MeganavTheme } from "./LegacyMeganav";
 
 type MeganavDesktopItems = {
   panels: MeganavPanels;
