+++ src/core/MeganavSearch.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import Icon from "./Icon";
 import MeganavSearchPanel from "./MeganavSearchPanel";
-import { AbsUrl } from "./Meganav";
+import { AbsUrl } from "./LegacyMeganav";
 
 const MeganavSearch = ({
   absUrl,
